requirements.txt
streamlit
python-dotenv

PyMuPDF
easyocr
torch
torchvision
numpy
Pillow

diffusers
transformers
accelerate

sentencepiece
gigachat
json_repair
requests